refactor: repository 삭제 및 테스트 코드 위치 변경
