+++ src/main/java/controller/BlackJackController.java
@@ -14,15 +14,12 @@ import dto.ParticipantsDto;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import repository.BetAmountRepository;
 import view.OutputView;
 
 public class BlackJackController {
 
     private final InputController inputController;
     private final OutputView outputView;
-//    private final BetAmountRepository repository = new BetAmountRepository(); // TODO 리팩터링 필요
-
 
     public BlackJackController(final InputController inputController, final OutputView outputView) {
         this.inputController = inputController;
@@ -37,7 +34,6 @@ public class BlackJackController {
         for (final Name name : names.getPlayerNames()) {
             BetAmount betAmount = inputController.getBetAmount(name.getValue());
             temp.add(new Player(name, betAmount));
-//            repository.put(player, betAmount);
         }
 
         final Players players = new Players(temp);
@@ -83,7 +79,6 @@ public class BlackJackController {
         Map<Player, Amount> finalResult = players.calculateResult(dealer);
         Amount dealerAmount = dealer.calculateRevenue(finalResult);
         outputView.printGameResult(finalResult, dealerAmount);
-//        outputView.printGameResult(repository.calculateResult(dealer), repository.calculateDealerAmount());
     }
 
     private void deal(final Player player, final Dealer dealer) {
